+++ CHANGELOG.md
@@ -1,2 +1,7 @@
+# 0.2.0
+* Add chargeback parser, CBNOTParser
+* Create new class for transaction parser, EPTRNParser
+
+
 # 0.1.0
 * Add type parsing, read from config file